doma-genをKotlin用にリメイク

+++ business/src/main/kotlin/com/isystk/sample/common/dto/mail/MailStockPaymentComplete.kt
@@ -5,5 +5,5 @@ package com.isystk.sample.common.dto.mail
  */
 class MailStockPaymentComplete {
     var userName: String? = null
-    var amount: Int? = null
+    var amount: Long? = null
 }
